+++ src/filehandler.h
@@ -1,6 +1,6 @@
 #ifndef FILEHANDLER_H
 #define FILEHANDLER_H
-
+#include <functional>   /*  uint8_t    */
 #include <vector>
 
 class BMPFilehandler {
@@ -17,6 +17,6 @@ class BMPFilehandler {
         void WriteBMP(int width, int height, std::vector<std::vector<unsigned int>>& data);
 };
 
-void writeBMP(char *outp_filename, int width, int height, unsigned char *header, std::vector<std::vector<unsigned int>>& data);
+void writeBMP(char *outp_filename, int width, int height, unsigned char *header, std::vector<std::vector<uint8_t>>& data);
 
 #endif
